+++ src/pages/planos/plano-negocios.tsx
@@ -4,6 +4,7 @@ import { graphql } from 'gatsby'
 import type { PlanoNegocioQuery } from '@generated/graphql'
 import { mark } from 'src/sdk/tests/mark'
 import Breadcrumb from 'src/components/sections/Breadcrumb'
+import BuyBox from 'src/components/sections/BuyBox/BuyBox'
 import ExplanationPlan from 'src/components/sections/ExplanationPlan'
 import type { SearchState } from '@faststore/sdk'
 import { SearchProvider, parseSearchState } from '@faststore/sdk'
@@ -75,6 +76,11 @@ function Page(props: Props) {
       {...searchParams}
     >
       <Breadcrumb breadcrumbList={itemListElement} name={title} />
+      <BuyBox
+        nodes={allContentfulPlanos.nodes.filter(
+          (node) => node.slug === '/plano-negocios'
+        )}
+      />
       <ExplanationPlan
         nodes={allContentfulPlanos.nodes.filter(
           (node) => node.slug === '/plano-negocios'
@@ -101,8 +107,22 @@ export const querySSG = graphql`
         titleTemplate
       }
     }
+
     allContentfulPlanos(sort: { order: ASC, fields: createdAt }) {
       nodes {
+        textoBotao
+        titulo
+        preco
+        slug
+        bannerImageMobile {
+          url
+        }
+        bannerImageDesktop {
+          url
+        }
+        compartilhar {
+          url
+        }
         galleryTitle
         texto {
           texto