+++ CHANGELOG.md
@@ -6,18 +6,21 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
 ### Added
+
 - **GM-16** - Component Mult Range Slider.
 - **GM-114** - Page plans Gallery filter.
 
 - **GM-113** - ExplanationPlan and Subtitle component.
+
 ### Fixed
 
 ## [0.8.0] - 2022-07-11
 
 ### Added
 
--**GM-123** - Created Created Banner component on for companies.
+- **GM-116** - Created Buy Box -**GM-123** - Created Created Banner component on for companies.
 
 - **GM-125** - Created Benefits Plans component business.
 - **GM-121** - Created Benefits Plans component.